# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  controllers.userBody:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      role_id:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  models.Role:
    properties:
      id:
        type: integer
      name:
        type: string
      permissions:
        items:
          type: string
        type: array
    type: object
...
  models.User:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      role:
        $ref: '#/definitions/models.Role'
      role_id:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
info:
  contact:
    name: THS Armada Development
    url: https://github.com/armada-ths
  description: REST API powering the THS Armada career fair website (armada.nu) and
    its admin panel.
  title: ArmadaCMS API
  version: "1.0"
paths:
...
  /customusers:
    get:
      parameters:
      - description: Pagination range, e.g. [0,24]
        in: query
        name: range
        type: string
      - description: Sort, e.g. [\
        in: query
        name: sort
        type: string
      - description: Filter, e.g. {\
        in: query
        name: filter
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Content-Range:
              description: users 0-24/100
              type: string
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      parameters:
      - description: User data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.userBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Invalid body
          schema:
            type: string
        "500":
          description: Create failed
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create user
      tags:
      - users
  /customusers/{id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Deleted
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - users
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "404":
          description: User not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated user data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.userBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Invalid data
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
        "500":
          description: Update failed
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - users
...
  /me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: User not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
...
  /roles:
    get:
      parameters:
      - description: Pagination range, e.g. [0,24]
        in: query
        name: range
        type: string
      - description: Sort, e.g. [\
        in: query
        name: sort
        type: string
      - description: Filter, e.g. {\
        in: query
        name: filter
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Content-Range:
              description: roles 0-24/10
              type: string
          schema:
            items:
              $ref: '#/definitions/models.Role'
            type: array
      security:
      - BearerAuth: []
      summary: List roles
      tags:
      - roles
    post:
      consumes:
      - application/json
      parameters:
      - description: Role data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Role'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Role'
        "400":
          description: Invalid body
          schema:
            type: string
        "500":
          description: Create failed
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create role
      tags:
      - roles
  /roles/{id}:
    delete:
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Deleted
          schema:
            type: string
        "404":
          description: Role not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete role
      tags:
      - roles
    get:
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Role'
        "404":
          description: Role not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get role by ID
      tags:
      - roles
    put:
      consumes:
      - application/json
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated role data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Role'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Role'
        "400":
          description: Invalid body
          schema:
            type: string
        "404":
          description: Role not found
          schema:
            type: string
        "500":
          description: Update failed
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update role
      tags:
      - roles
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer JWT token obtained from POST /api/v1/login. Format: "Bearer
      <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"